requirements.txt
openai
streamlit
chromadb
pandas
python-dotenv
langchain
sentence-transformers
